backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
google-auth
requests
